v5.x-refactor-tests: removed Evaluation trait, CostAccumulatorTests.scala
